Add drawing based on hardcoded directory, clamping nodes and file size printing

+++ headers/Node.hpp
@@ -10,7 +10,7 @@ namespace dv
     class Node
     {
         public:
-            Node(const std::string& name, unsigned int id, EConnectionType connT, EFileType fType, unsigned int size = 4096);
+            Node(const std::string& name, unsigned int id, EConnectionType connT, EFileType fType, long size = 4096);
             ~Node();
             void Reparent(Node* newParent);
             void RemoveChild(Node* child);
@@ -21,7 +21,7 @@ namespace dv
             EConnectionType ConnnectionType; 
             EFileType Type;
             //permissons?
-            unsigned int Size; //default size is 4096 for a directory(in bytes) //may be bigger than int?
+            long Size; //default size is 4096 for a directory(in bytes) //may be bigger than int?
             Node* Parent = nullptr;
             std::vector<Node*> Children;
     };
